remake server

+++ src/modules/books/infra/prisma/repositories/BooksPrismaRepository.ts
@@ -1,6 +1,10 @@
+import { inject, injectable } from 'tsyringe'
+
 import { type IUpdateBookDTO } from '@modules/books/dtos/IUpdateBookDTO'
-import { type IUpdateFrontCoverDTO } from '@modules/books/dtos/IUpdateFrontCoverDTO'
 import { type Prisma } from '@prisma/client'
+import { ICacheProvider } from '@shared/container/providers/CacheProvider/ICacheProvider'
+import InjectableDependencies from '@shared/container/types'
+import { makeErrorBookNotFound } from '@shared/errors/books/makeErrorBookNotFound'
 import { prisma } from '@shared/infra/database/createConnection'
 
 import { type IBooksRepository } from '../../repositories/contracts/IBooksRepository'
@@ -8,7 +12,7 @@ import { type IBook } from '../../repositories/entities/IBook'
 import { type ICapitule } from '../../repositories/entities/ICapitule'
 import { type IGenre } from '../../repositories/entities/IGenre'
 
-const defaultInclude: Prisma.BookInclude = {
+export const defaultIncludeBook: Prisma.BookInclude = {
   genres: true,
   authors: {
     select: {
@@ -40,16 +44,56 @@ const defaultInclude: Prisma.BookInclude = {
         },
       },
     },
+    orderBy: {
+      sequence: 'asc',
+    },
   },
   comments: true,
 }
 
+@injectable()
 export class BooksPrismaRepository implements IBooksRepository {
+  constructor(
+    @inject(InjectableDependencies.Providers.CacheProvider)
+    private readonly cacheProvider: ICacheProvider,
+  ) {}
+
+  private async setBookInCache(book: IBook): Promise<void> {
+    await this.cacheProvider.setInfo<IBook>(
+      {
+        key: 'book',
+        objectId: book.id,
+      },
+      book,
+      60 * 60 * 24, // 1 day
+    )
+  }
+
+  private async getBookInCache(bookId: string): Promise<IBook | null> {
+    return await this.cacheProvider.getInfo<IBook>({
+      key: 'book',
+      objectId: bookId,
+    })
+  }
+
   async create(data: Prisma.BookUncheckedCreateInput): Promise<IBook | null> {
     const book = await prisma.book.create({
       data,
-      include: defaultInclude,
+      include: defaultIncludeBook,
     })
+
+    if (book) {
+      Promise.all([
+        this.setBookInCache(book),
+
+        this.cacheProvider.delete({
+          key: 'project',
+          objectId: book.project_id,
+        }),
+      ]).catch((err) => {
+        throw err
+      })
+    }
 
     return book
   }
@@ -61,26 +105,77 @@ export class BooksPrismaRepository implements IBooksRepository {
       data,
     })
 
+    this.cacheProvider
+      .delete({
+        key: 'book',
+        objectId: '*',
+      })
+      .catch((err) => {
+        throw err
+      })
+
     return genre
   }
 
   async findById(id: string): Promise<IBook | null> {
+    const bookInCache = await this.getBookInCache(id)
+    if (bookInCache) return bookInCache
+
     const book = await prisma.book.findUnique({
       where: {
         id,
       },
-      include: defaultInclude,
+      include: defaultIncludeBook,
     })
+
+    if (book) {
+      this.setBookInCache(book).catch((err) => {
+        throw err
+      })
+    }
 
     return book
   }
 
   async delete(id: string): Promise<void> {
-    await prisma.book.delete({
+    const book = await prisma.book.findUnique({
       where: {
         id,
       },
+      select: {
+        project_id: true,
+        capitules: {
+          select: {
+            id: true,
+          },
+        },
+      },
     })
+    if (!book) throw makeErrorBookNotFound()
+
+    const promises = [
+      prisma.book.delete({
+        where: {
+          id,
+        },
+      }),
+      this.cacheProvider.delete({
+        key: 'project',
+        objectId: book.project_id,
+      }),
+      this.cacheProvider.delete({
+        key: 'book',
+        objectId: id,
+      }),
+    ]
+
+    book.capitules.map((capitule) =>
+      promises.push(
+        this.cacheProvider.delete({ key: 'capitule', objectId: capitule.id }),
+      ),
+    )
+
+    await Promise.all(promises)
   }
 
   async update({ bookId, data }: IUpdateBookDTO): Promise<IBook | null> {
@@ -89,41 +184,42 @@ export class BooksPrismaRepository implements IBooksRepository {
         id: bookId,
       },
       data,
-      include: defaultInclude,
+      include: defaultIncludeBook,
     })
+
+    if (book) {
+      Promise.all([
+        this.setBookInCache(book),
+
+        this.cacheProvider.delete({
+          key: 'project',
+          objectId: book.project_id,
+        }),
+      ]).catch((err) => {
+        throw err
+      })
+    }
 
     return book
   }
 
-  async updateFrontCover({
-    bookId,
-    front_cover_filename,
-    front_cover_url,
-  }: IUpdateFrontCoverDTO): Promise<void> {
-    await prisma.book.update({
-      where: {
-        id: bookId,
-      },
-      data: {
-        front_cover_filename,
-        front_cover_url,
-      },
-    })
-  }
-
   async removeGenreOfBook(genreId: string, bookId: string): Promise<void> {
-    await prisma.genre.update({
-      where: {
-        id: genreId,
-      },
-      data: {
-        books: {
-          delete: {
-            id: bookId,
+    await Promise.all([
+      prisma.genre.update({
+        where: {
+          id: genreId,
+        },
+        data: {
+          books: {
+            disconnect: {
+              id: bookId,
+            },
           },
         },
-      },
-    })
+      }),
+
+      this.cacheProvider.delete({ key: 'book', objectId: bookId }),
+    ])
   }
 
   async listCapitules(bookId: string): Promise<ICapitule[]> {
