+++ src/modules/persons/infra/prisma/repositories/CouplesPrismaRepository.ts
@@ -1,11 +1,22 @@
+import { inject, injectable } from 'tsyringe'
+
 import { type ICreateCoupleDTO } from '@modules/persons/dtos/ICreateCoupleDTO'
 import { type IUpdateCoupleDTO } from '@modules/persons/dtos/IUpdateCoupleDTO'
+import { ICacheProvider } from '@shared/container/providers/CacheProvider/ICacheProvider'
+import InjectableDependencies from '@shared/container/types'
+import { makeErrorNotFound } from '@shared/errors/useFull/makeErrorNotFound'
 import { prisma } from '@shared/infra/database/createConnection'
 
 import { type ICouplesRepository } from '../../repositories/contracts/ICouplesRepository'
 import { type ICouple } from '../../repositories/entities/ICouple'
 
+@injectable()
 export class CouplesPrismaRepository implements ICouplesRepository {
+  constructor(
+    @inject(InjectableDependencies.Providers.CacheProvider)
+    private readonly cacheProvider: ICacheProvider,
+  ) {}
+
   async create(data: ICreateCoupleDTO): Promise<ICouple | null> {
     const couple = await prisma.couple.create({
       data,
@@ -14,14 +25,59 @@ export class CouplesPrismaRepository implements ICouplesRepository {
       },
     })
 
+    if (couple) {
+      Promise.all([
+        this.cacheProvider.delete({
+          key: 'person',
+          objectId: couple.person_id,
+        }),
+
+        this.cacheProvider.delete({
+          key: 'person',
+          objectId: couple.coupleWithPerson.person_id,
+        }),
+      ]).catch((err) => {
+        throw err
+      })
+    }
+
     return couple
   }
 
   async delete(coupleId: string): Promise<void> {
-    await prisma.couple.delete({
+    const couple = await prisma.couple.findUnique({
       where: {
         id: coupleId,
       },
+      select: {
+        person_id: true,
+        coupleWithPerson: {
+          select: {
+            person_id: true,
+          },
+        },
+      },
+    })
+    if (!couple) throw makeErrorNotFound({ whatsNotFound: 'Casal' })
+
+    const promises = [
+      prisma.couple.delete({
+        where: {
+          id: coupleId,
+        },
+      }),
+      this.cacheProvider.delete({
+        key: 'person',
+        objectId: couple.coupleWithPerson.person_id,
+      }),
+      this.cacheProvider.delete({
+        key: 'person',
+        objectId: couple.person_id,
+      }),
+    ]
+
+    await Promise.all(promises).catch((err) => {
+      throw err
     })
   }
 
@@ -44,6 +100,21 @@ export class CouplesPrismaRepository implements ICouplesRepository {
       include: {
         coupleWithPerson: true,
       },
+    })
+
+    const promises = [
+      this.cacheProvider.delete({
+        key: 'person',
+        objectId: couple.coupleWithPerson.person_id,
+      }),
+      this.cacheProvider.delete({
+        key: 'person',
+        objectId: couple.person_id,
+      }),
+    ]
+
+    Promise.all(promises).catch((err) => {
+      throw err
     })
 
     return couple
