+++ src/modules/persons/infra/prisma/repositories/DreamsPrismaRepository.ts
@@ -1,6 +1,11 @@
+import { inject, injectable } from 'tsyringe'
+
 import { type ICreateDreamDTO } from '@modules/persons/dtos/ICreateDreamDTO'
 import { type IUpdateDreamDTO } from '@modules/persons/dtos/IUpdateDreamDTO'
-import { type Dream } from '@prisma/client'
+import { type Prisma, type Dream } from '@prisma/client'
+import { ICacheProvider } from '@shared/container/providers/CacheProvider/ICacheProvider'
+import InjectableDependencies from '@shared/container/types'
+import { makeErrorNotFound } from '@shared/errors/useFull/makeErrorNotFound'
 import { prisma } from '@shared/infra/database/createConnection'
 
 import { type IDreamsRepository } from '../../repositories/contracts/IDreamsRepository'
@@ -8,11 +13,28 @@ import { type IDream } from '../../repositories/entities/IDream'
 import { type IAddOnePersonInObject } from '../../repositories/types/IAddOnePersonInObject'
 import { type IRemoveOnePersonById } from '../../repositories/types/IRemoveOnePersonById'
 
+@injectable()
 export class DreamsPrismaRepository implements IDreamsRepository {
-  async create(data: ICreateDreamDTO): Promise<Dream | null> {
+  constructor(
+    @inject(InjectableDependencies.Providers.CacheProvider)
+    private readonly cacheProvider: ICacheProvider,
+  ) {}
+
+  async create(data: ICreateDreamDTO, personId: string): Promise<Dream | null> {
     const dream = await prisma.dream.create({
       data,
     })
+
+    if (dream) {
+      this.cacheProvider
+        .delete({
+          key: 'person',
+          objectId: personId,
+        })
+        .catch((err) => {
+          throw err
+        })
+    }
 
     return dream
   }
@@ -25,6 +47,8 @@ export class DreamsPrismaRepository implements IDreamsRepository {
       include: {
         persons: {
           select: {
+            image_url: true,
+            name: true,
             id: true,
           },
         },
@@ -35,10 +59,43 @@ export class DreamsPrismaRepository implements IDreamsRepository {
   }
 
   async delete(dreamId: string): Promise<void> {
-    await prisma.dream.delete({
+    const dream = await prisma.dream.findUnique({
       where: {
         id: dreamId,
       },
+      select: {
+        persons: {
+          select: {
+            id: true,
+          },
+        },
+      },
+    })
+    if (!dream) throw makeErrorNotFound({ whatsNotFound: 'Sonho' })
+
+    const personsToCleanInCache = dream.persons.map((person) => person.id)
+
+    const promises: Array<
+      Promise<void> | Prisma.Prisma__DreamClient<Dream, never>
+    > = [
+      prisma.dream.delete({
+        where: {
+          id: dreamId,
+        },
+      }),
+    ]
+
+    personsToCleanInCache.map((personToCleanInCache) =>
+      promises.push(
+        this.cacheProvider.delete({
+          key: 'person',
+          objectId: personToCleanInCache,
+        }),
+      ),
+    )
+
+    await Promise.all(promises).catch((err) => {
+      throw err
     })
   }
 
@@ -46,17 +103,22 @@ export class DreamsPrismaRepository implements IDreamsRepository {
     objectId,
     personId,
   }: IRemoveOnePersonById): Promise<void> {
-    await prisma.dream.update({
-      where: {
-        id: objectId,
-      },
-      data: {
-        persons: {
-          delete: {
-            id: personId,
+    await Promise.all([
+      prisma.dream.update({
+        where: {
+          id: objectId,
+        },
+        data: {
+          persons: {
+            disconnect: {
+              id: personId,
+            },
           },
         },
-      },
+      }),
+      this.cacheProvider.delete({ key: 'person', objectId: personId }),
+    ]).catch((err) => {
+      throw err
     })
   }
 
@@ -64,17 +126,22 @@ export class DreamsPrismaRepository implements IDreamsRepository {
     objectId,
     personId,
   }: IAddOnePersonInObject): Promise<void> {
-    await prisma.dream.update({
-      where: {
-        id: objectId,
-      },
-      data: {
-        persons: {
-          connect: {
-            id: personId,
+    await Promise.all([
+      prisma.dream.update({
+        where: {
+          id: objectId,
+        },
+        data: {
+          persons: {
+            connect: {
+              id: personId,
+            },
           },
         },
-      },
+      }),
+      this.cacheProvider.delete({ key: 'person', objectId: personId }),
+    ]).catch((err) => {
+      throw err
     })
   }
 
@@ -84,8 +151,57 @@ export class DreamsPrismaRepository implements IDreamsRepository {
         id: dreamId,
       },
       data,
+      include: {
+        persons: {
+          select: {
+            id: true,
+            name: true,
+            image_url: true,
+          },
+        },
+      },
+    })
+
+    if (!dream) throw makeErrorNotFound({ whatsNotFound: 'Sonho' })
+
+    const personsToCleanInCache = dream.persons.map((person) => person.id)
+
+    Promise.all(
+      personsToCleanInCache.map(async (personToCleanInCache) => {
+        await this.cacheProvider.delete({
+          key: 'person',
+          objectId: personToCleanInCache,
+        })
+      }),
+    ).catch((err) => {
+      throw err
     })
 
     return dream
+  }
+
+  async listPerPersons(personIds: string[]): Promise<IDream[]> {
+    const dreams = await prisma.dream.findMany({
+      where: {
+        persons: {
+          some: {
+            id: {
+              in: personIds,
+            },
+          },
+        },
+      },
+      include: {
+        persons: {
+          select: {
+            image_url: true,
+            name: true,
+            id: true,
+          },
+        },
+      },
+    })
+
+    return dreams
   }
 }
